compile rstime

+++ db/repl/rstime.h
@@ -6,8 +6,8 @@ namespace mongo {
 
     struct rstime { 
         unsigned long long v;
-        unsigned server() const { return v >> (64-8); }
-        unsigned op() const { return v & 0xffffffffffffffL; }
+        unsigned server() const { return (unsigned) (v >> (64-8)); }
+        unsigned long long op() const { return v & 0xffffffffffffffL; }
     };
 
 }
